Add files via upload
Exploratory Data Analysis of the Thought-Provoking YouTube Channels


In this era, Youtube is not only the second most searched engine in this world but the largest school in the world. One of the best platforms for learning and growing. Since its foundation in 2005, content creation has reached its peak. People, in every minute, upload more than 500 hours of videos and making it the most streamed platform, overflowing with tonnes of information generated.

As I am a seeker of knowledge and thought-provoking ideas, I find it the most intelligent platform to gain insights. Hovering around this school made me find intellectual channels that helped me in self-actualization and improvement in all sectors of my life.

Being new in the data field, I thought of doing an analysis on these thought-provoking channels to gain insights and analyze how this youtube video system works. There has been a lot of confusion among the audience about the reach of the videos, and what their success depends on.

It might be helpful for a new audience that is trying to engage in this business, to know what has the most impact on the success of the videos, whether it's duration length, views, likes, or comments and the relation between them.

This is my portfolio project and is limited to the channels of my interests. Therefore, in this project, I will perform the statistical analysis of around 6 intellectual and thought-provoking Youtube channels.